refactor: Alteração de erros 400 para 422 que são mais específicos
