+++ api/src/main/java/com/dbserver/votingchallenge/exceptions/associated/AssociatedNotFoundException.java
@@ -4,6 +4,6 @@ import com.dbserver.votingchallenge.exceptions.general.NotFoundException;
 
 public class AssociatedNotFoundException extends NotFoundException {
     public AssociatedNotFoundException() {
-        super("Associado não encontrado");
+        super("Usuário associado não encontrado");
     }
 }
